compose: Move call button display functions to compose_call_ui.
