+++ web/src/compose.js
@@ -8,7 +8,6 @@ import render_success_message_scheduled_banner from "../templates/compose_banner
 
 import * as channel from "./channel";
 import * as compose_banner from "./compose_banner";
-import * as compose_call from "./compose_call";
 import * as compose_state from "./compose_state";
 import * as compose_ui from "./compose_ui";
 import * as compose_validate from "./compose_validate";
@@ -30,23 +29,6 @@ import * as util from "./util";
 import * as zcommand from "./zcommand";
 
 // Docs: https://zulip.readthedocs.io/en/latest/subsystems/sending-messages.html
-
-export function update_audio_and_video_chat_button_display() {
-    update_audio_chat_button_display();
-    update_video_chat_button_display();
-}
-
-export function update_video_chat_button_display() {
-    const show_video_chat_button = compose_call.compute_show_video_chat_button();
-    $("#below-compose-content .video_link").toggle(show_video_chat_button);
-    $(".message-edit-feature-group .video_link").toggle(show_video_chat_button);
-}
-
-export function update_audio_chat_button_display() {
-    const show_audio_chat_button = compose_call.compute_show_audio_chat_button();
-    $("#below-compose-content .audio_link").toggle(show_audio_chat_button);
-    $(".message-edit-feature-group .audio_link").toggle(show_audio_chat_button);
-}
 
 export function clear_invites() {
     $(
